+++ pkg/server/BUILD.bazel
@@ -48,12 +48,9 @@ go_library(
         "rlimit_unix.go",
         "server.go",
         "server_controller.go",
-        "server_controller_accessors.go",
         "server_controller_channel_orchestrator.go",
         "server_controller_http.go",
         "server_controller_new_server.go",
-        "server_controller_orchestration.go",
-        "server_controller_orchestration_method.go",
         "server_controller_sql.go",
         "server_http.go",
         "server_obs_service.go",
