+++ pkg/server/server_controller_channel_orchestrator.go
@@ -33,7 +33,11 @@ type channelOrchestrator struct {
 	serverFactory serverFactoryForOrchestration
 }
 
-var _ serverOrchestrator = (*channelOrchestrator)(nil)
+// serverFactoryForOrchestration provides the method that instantiates tenant servers.
+type serverFactoryForOrchestration interface {
+	// newServerForOrchestrator returns a new tenant server.
+	newServerForOrchestrator(ctx context.Context, nc *roachpb.TenantNameContainer, tenantStopper *stop.Stopper) (orchestratedServer, error)
+}
 
 func newChannelOrchestrator(
 	parentStopper *stop.Stopper, serverFactory serverFactoryForOrchestration,
@@ -44,8 +48,9 @@ func newChannelOrchestrator(
 	}
 }
 
-// serverStateUsingChannels coordinates the lifecycle of a tenant
-// server. It ensures sane concurrent behavior between:
+// serverState coordinates the lifecycle of a tenant server. It
+// ensures sane concurrent behavior between:
+//
 // - requests to start a server manually, e.g. via testServer;
 // - async changes to the tenant service mode;
 // - quiescence of the outer stopper;
@@ -67,7 +72,7 @@ func newChannelOrchestrator(
 //
 // The async task is also responsible for reporting the server
 // start/stop events in the event log.
-type serverStateUsingChannels struct {
+type serverState struct {
 	// nc holds a shared reference to the current name of the
 	// tenant. If the tenant's name is updated, the `Set` method on
 	// nameContainer should be called in order to update any subscribers
@@ -105,51 +110,49 @@ type serverStateUsingChannels struct {
 	stoppedCh <-chan struct{}
 }
 
-var _ serverState = (*serverStateUsingChannels)(nil)
-
 // getServer is part of the serverState interface.
-func (s *serverStateUsingChannels) getServer() (orchestratedServer, bool) {
+func (s *serverState) getServer() (orchestratedServer, bool) {
 	s.startedMu.Lock()
 	defer s.startedMu.Unlock()
 	return s.startedMu.server, s.startedMu.server != nil
 }
 
 // nameContainer is part of the serverState interface.
-func (s *serverStateUsingChannels) nameContainer() *roachpb.TenantNameContainer {
+func (s *serverState) nameContainer() *roachpb.TenantNameContainer {
 	return s.nc
 }
 
 // getLastStartupError is part of the serverState interface.
-func (s *serverStateUsingChannels) getLastStartupError() error {
+func (s *serverState) getLastStartupError() error {
 	return s.startErr
 }
 
 // requestGracefulShutdown is part of the serverState interface.
-func (s *serverStateUsingChannels) requestGracefulShutdown(ctx context.Context) {
+func (s *serverState) requestGracefulShutdown(ctx context.Context) {
 	// TODO(knz): this is incorrect because it does not obey the
 	// incoming context's cancellation.
 	s.requestGracefulStop()
 }
 
 // requestImmediateShutdown is part of the serverState interface.
-func (s *serverStateUsingChannels) requestImmediateShutdown(ctx context.Context) {
+func (s *serverState) requestImmediateShutdown(ctx context.Context) {
 	s.requestImmediateStop()
 }
 
 // stopped is part of the serverState interface.
-func (s *serverStateUsingChannels) stopped() <-chan struct{} {
+func (s *serverState) stopped() <-chan struct{} {
 	return s.stoppedCh
 }
 
 // startedOrStopped is part of the serverState interface.
-func (s *serverStateUsingChannels) startedOrStopped() <-chan struct{} {
+func (s *serverState) startedOrStopped() <-chan struct{} {
 	return s.startedOrStoppedCh
 }
 
 // makeServerStateForSystemTenant is part of the orchestrator interface.
 func (o *channelOrchestrator) makeServerStateForSystemTenant(
 	nc *roachpb.TenantNameContainer, systemSrv orchestratedServer,
-) serverState {
+) *serverState {
 	// We make the serverState for the system mock the regular
 	// lifecycle. It starts with an already-closed `startedOrStopped`
 	// channel; and it properly reacts to a call to requestStop()
@@ -157,7 +160,7 @@ func (o *channelOrchestrator) makeServerStateForSystemTenant(
 	closedChan := make(chan struct{})
 	close(closedChan)
 	closeCtx, cancelFn := context.WithCancel(context.Background())
-	st := &serverStateUsingChannels{
+	st := &serverState{
 		nc:                   nc,
 		startedOrStoppedCh:   closedChan,
 		requestImmediateStop: cancelFn,
@@ -192,7 +195,7 @@ func (o *channelOrchestrator) startControlledServer(
 	// synchronize on the server's state, use the resulting
 	// serverState instead.
 	serverStoppingFn func(ctx context.Context, tenantName roachpb.TenantName, tid roachpb.TenantID, sid base.SQLInstanceID),
-) (serverState, error) {
+) (*serverState, error) {
 	var immediateStopRequest sync.Once
 	immediateStopRequestCh := make(chan struct{})
 	immediateStopFn := func() {
@@ -211,7 +214,7 @@ func (o *channelOrchestrator) startControlledServer(
 	stoppedCh := make(chan struct{})
 	startedOrStoppedCh := make(chan struct{})
 
-	state := &serverStateUsingChannels{
+	state := &serverState{
 		nc:                   roachpb.NewTenantNameContainer(tenantName),
 		startedOrStoppedCh:   startedOrStoppedCh,
 		requestImmediateStop: immediateStopFn,
